requirements.txt
# Core
typer[all]>=0.9.0
loguru>=0.7.2
pyyaml>=6.0.1
rich
questionary

# Database
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0

# Networking & Scraping
aiohttp>=3.9.1
aiohttp-socks>=0.8.0
beautifulsoup4>=4.12.2
aiofiles>=23.2.1